# loaderCache: copy map spec before using it